Need to make it a double primitive rather than Double to cast to an int

+++ src/main/java/frc/robot/vision/CameraSwitcher.java
@@ -120,6 +120,6 @@ public class CameraSwitcher {
   }
 
   public int totalCameras() {
-    return (int) totalNetworkTableEntry.getNumber(4);
+    return (int) totalNetworkTableEntry.getNumber(4).doubleValue();
   }
 }
